refactor: restructing order of submodule impls
