+++ src/components/dh_cartridge/mirroring.rs
@@ -1,0 +1,7 @@
+#[derive(Debug)]
+pub enum Mirroring {
+    Horizontal,
+    Vertical,
+    OnescreanrLlo,
+    OnescreanHi,
+}
