+++ src/components/dh_mappers/m000/impls/defaut.rs
@@ -1,0 +1,13 @@
+use std::marker::PhantomData;
+
+use crate::components::dh_mappers::m000::types::M000;
+
+// impl Default for M000 {
+//     fn default() -> Self {
+//         M000 {
+//             _marker: PhantomData,
+//             chr_bank: 0,
+//             prg_bank: 0,
+//         }
+//     }
+// }
