+++ src/components/dh_mappers/mapper_data/types.rs
@@ -1,0 +1,9 @@
+use std::marker::PhantomData;
+
+// trivially copiable struct on most systems: mov QWORD
+#[derive(Debug, Clone, Copy)]
+pub struct MapperData<MapperType: Default> {
+    pub _marker: PhantomData<MapperType>,
+    pub chr_bank: u8,
+    pub prg_bank: u8,
+}
