--- src/components/mappers/mod.rs
@@ -1,2 +1,0 @@
-pub mod m000;
-pub mod mapper;
